+++ packages/cli/lib/cli.js
@@ -1,4 +1,6 @@
-import { program } from 'commander'
+import { CommanderError, program } from 'commander'
+import isInstalledGlobally from 'is-installed-globally'
+import { MultipleRootsError } from '../findPipeline.js'
 import runAction from './cli/runAction.js'
 import * as monitoringOptions from './cli/monitoringOptions.js'
 import * as commonOptions from './cli/commonOptions.js'
@@ -24,6 +26,7 @@ import { combine } from './cli/options.js'
  */
 
 program
+  .name('barnard59')
   .addOption(commonOptions.variable)
   .addOption(commonOptions.variableAll)
   .addOption(commonOptions.verbose)
@@ -36,7 +39,7 @@ program
 
 const runCommand = program
   .command('run <filename>')
-  .option('--output [filename]', 'output file', '-')
+  .option('--output [filename]', 'Output file', '-')
   .option('--pipeline [iri]', 'IRI of the pipeline description')
   .action(async (filename, options) => {
     const combinedOptions = combine(options)
@@ -58,5 +61,34 @@ export default async function () {
     .addOption(commonOptions.verbose)
     .addOption(commonOptions.quiet)
 
-  await program.parseAsync(process.argv)
+  return {
+    program,
+    async run() {
+      program.exitOverride()
+
+      try {
+        await program.parseAsync(process.argv)
+      } catch (/** @type {any} */ error) {
+        if (error instanceof CommanderError) {
+          const { groups } = /unknown command '(?<command>[^']+)'/.exec(error.message) || {}
+          if (groups && groups.command) {
+            /* eslint-disable no-console */
+            if (isInstalledGlobally) {
+              console.error(`Try running 'npm install (-g) barnard59-${groups.command}'`)
+            }
+
+            console.error(`Try running 'npm install barnard59-${groups.command}'`)
+          }
+          process.exit(error.exitCode)
+        } else if (error instanceof MultipleRootsError) {
+          const alternatives = error.alternatives.map(x => `\n\t--pipeline ${x}`).join('')
+          // eslint-disable-next-line no-console
+          console.error(`Multiple root pipelines found. Try one of these:${alternatives}`)
+          process.exit(1)
+        } else {
+          throw error
+        }
+      }
+    },
+  }
 }