+++ test/e2e/CHANGELOG.md
@@ -1,5 +1,20 @@
 # barnard59-test-e2e
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [01682f7]
+  - barnard59-formats@4.0.0
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [8282b0e]
+- Updated dependencies [8282b0e]
+  - barnard59-formats@3.0.0
+
 ## 0.1.1
 
 ### Patch Changes